fix(pvcm): rename plugin to pv_mcu.so to match cont_ctrl type string
